Update README.md
At this project, it's possible to adjust the time of the fish feed during the day. 
It's was used as a micro servo 90g attached at a 3d printed enclosure, an Arduino nano, and a RTC 3231 to provide the time reference if the main power goes down and the controller is a reboot.